config: split up config.c modules
